requirements.txt
py-require==0.17
nr==1.4.0
